+++ src/main/java/net/sf/jabref/gui/worker/MarkEntriesAction.java
@@ -18,9 +18,6 @@ import net.sf.jabref.preferences.JabRefPreferences;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-/**
- *
- */
 public class MarkEntriesAction extends AbstractWorker implements ActionListener {
 
     private final JabRefFrame frame;